Moving the creation of config files to run time instead installation time
Moving the creation of config files to run time instead installation time

Fix template error in installation

An error was raised when kytos was trying to copy template files to installation root folder. This commit changes config.py to solve this problem.
